fix test props
