+++ packages/app-harness/src/components/Pressable/index.tsx
@@ -1,17 +1,7 @@
 import React from 'react';
 import { Pressable, PressableProps } from '@flexn/create';
 import { useDebugContext } from '../../context/debugContext';
-import { isPlatformAndroid, isPlatformAndroidtv } from '@rnv/renative';
-
-export function testProps(testID: string | undefined) {
-    if (!testID) {
-        return;
-    }
-    if (isPlatformAndroid || isPlatformAndroidtv) {
-        return { accessibilityLabel: testID, accessible: true };
-    }
-    return { testID };
-}
+import { testProps } from '../../utils';
 
 const PressableEl = (props: PressableProps) => {
     const { setFocusedElementId } = useDebugContext();
